Files Description
pub.py - This is my [publisher code
rec.py - This is my receiver code. Here when I initially inserted I got a error that "exceeded the rate limit for object mutation operations (create, update, and delete). Please reduce your request rate." So I have inserted the data in scheduled intervals.
recz.py - This is the receiver code to compress the files and then store it into bucket.
rece.py - This is to encrypt the file, compress and store it in bucket.
dec.py - This is for decoding and decompress the files.
rsa.py - This is to generate private key and public key for encryption process